refactor(orulo): move fetch job to re_integrations
* Add JobQueue application on re_integrations depnedencies
to use JobQueue for get new buildings from orulo API.
* Create new FetchJobQueue on re_integrations umbrella.
* Add re_integrations migrations on formatter.
