+++ apps/re_integrations/.formatter.exs
@@ -1,5 +1,5 @@
 [
-  inputs: ["mix.exs", "lib/**/*.{ex,exs}"],
+  inputs: ["mix.exs", "lib/**/*.{ex,exs}", "priv/repo/migrations/*.exs"],
   locals_without_parens: [
     field: 2,
     field: 3
