+++ apps/re_integrations/lib/importers/orulo/building.ex
@@ -1,4 +1,4 @@
-defmodule Re.Importers.Orulo.Building do
+defmodule ReIntegrations.Importers.Orulo.Building do
   @moduledoc """
   Schema for Orulo Buildings sincronization
   """
